Revert "Projects page redesign (#4)"
This reverts commit 2785f05da85d8bf901ad790cd2857830aed2c9f1.
