--- nextjs/src/components/Projects/Disclaimer.tsx
@@ -1,10 +1,0 @@
-export default function Disclaimer() {
-    return (
-        <p className="-mb-44 text-xs sm:text-sm my-32 text-neutral-600 dark:text-neutral-300 text-center sm:max-w-5xl sm:mx-auto mx-5">
-            All logos and trademarks displayed on this page are the property of their respective
-            owners. The use of these logos is purely for informational purposes, to identify the
-            technologies used in these projects. The use of these logos does not imply any
-            endorsement or affiliation.
-        </p>
-    );
-}
