--- nextjs/src/components/Projects/OverlayCopy.tsx
@@ -1,33 +1,0 @@
-"use client";
-import { motion, useScroll, useTransform } from "framer-motion";
-import { useRef } from "react";
-
-interface OverlayCopyProps {
-    subheading: string;
-    heading: string;
-}
-
-export default function OverlayCopy({ subheading, heading }: OverlayCopyProps) {
-    const targetRef = useRef(null);
-    const { scrollYProgress } = useScroll({
-        target: targetRef,
-        offset: ["start end", "end start"]
-    });
-
-    const y = useTransform(scrollYProgress, [0, 1], [250, -250]);
-    const opacity = useTransform(scrollYProgress, [0.25, 0.5, 0.75], [0, 1, 0]);
-
-    return (
-        <motion.div
-            style={{
-                y,
-                opacity
-            }}
-            ref={targetRef}
-            className="absolute left-0 top-0 flex h-screen w-full flex-col items-center justify-center text-white"
-        >
-            <p className="mb-2 text-center text-xl md:mb-4 md:text-3xl">{subheading}</p>
-            <p className="text-center text-4xl font-bold md:text-7xl">{heading}</p>
-        </motion.div>
-    );
-}
