+++ nextjs/src/misc.ts
@@ -1,0 +1,77 @@
+export const frontendSkills = [
+    "React",
+    "TypeScript",
+    "JavaScript",
+    "TailwindCSS",
+    "React Query",
+    "Redux (RTK)",
+    "SEO",
+    "Next.js",
+    "Framer Motion",
+    "Firebase Auth",
+    "UI/UX Design",
+    "Playwright",
+    "Cypress",
+    "Jest",
+    "Vitest",
+    "HTML",
+    "CSS"
+];
+
+export const backendSkills = [
+    "TypeScript",
+    "Node.js",
+    "Express",
+    "Fastify",
+    "Prisma",
+    "PostgreSQL",
+    "SQLServer",
+    "MySQL",
+    "SQLite",
+    "MongoDB",
+    "REST API",
+    "GraphQL",
+    "JWT",
+    "Firebase Auth",
+    "Hoppscotch",
+    "Postman"
+];
+
+export const devOpsSkills = [
+    "Docker",
+    "Linux",
+    "CI/CD",
+    "GitHub",
+    "GitLab",
+    // "Kubernetes",
+    "NGINX",
+    "Bash",
+    "PowerShell",
+    "Python",
+    "Security",
+    "AWS",
+    "Firebase",
+    "Google Cloud",
+    "Oracle Cloud",
+    "Server management",
+    "Domain & SSL management"
+];
+
+export const otherSkills = [
+    "Agile",
+    "Scrum",
+    "Figma",
+    "Leadership",
+    "Communication",
+    "Building PCs",
+    "C++",
+    "C#",
+    "Rust",
+    "Java",
+    "Flutter",
+    "Dart",
+    "Troubleshooting",
+    "Team Management",
+    "Jetpack Compose",
+    "React Native"
+];
