--- nextjs/src/misc.tsx
@@ -1,129 +1,0 @@
-import { FaReact, FaNodeJs, FaCss3Alt } from "react-icons/fa";
-import {
-    SiCloudflare,
-    SiCypress,
-    SiDirectus,
-    SiDocker,
-    SiExpress,
-    SiFastify,
-    SiFirebase,
-    SiFramer,
-    SiJavascript,
-    SiMapbox,
-    SiMicrosoftsqlserver,
-    SiMongodb,
-    SiNextdotjs,
-    SiNginx,
-    SiOracle,
-    SiPlaywright,
-    SiPostgresql,
-    SiPrisma,
-    SiReactquery,
-    SiTailwindcss,
-    SiTypescript
-} from "react-icons/si";
-
-export const iconMapping: Record<string, JSX.Element> = {
-    React: <FaReact />,
-    "Node.js": <FaNodeJs />,
-    CSS3: <FaCss3Alt />,
-    "Next.js": <SiNextdotjs />,
-    TailwindCSS: <SiTailwindcss />,
-    JavaScript: <SiJavascript />,
-    TypeScript: <SiTypescript />,
-    "Firebase Auth": <SiFirebase />,
-    Prisma: <SiPrisma />,
-    Docker: <SiDocker />,
-    Fastify: <SiFastify />,
-    "SQL Server": <SiMicrosoftsqlserver />,
-    "Postgres SQL": <SiPostgresql />,
-    "Express.js": <SiExpress />,
-    MongoDB: <SiMongodb />,
-    "Framer Motion": <SiFramer />,
-    Directus: <SiDirectus />,
-    "Oracle Cloud": <SiOracle />,
-    Cloudflare: <SiCloudflare />,
-    NGINX: <SiNginx />,
-    "React Query": <SiReactquery />,
-    Mapbox: <SiMapbox />,
-    Playwright: <SiPlaywright />,
-    Cypress: <SiCypress />
-};
-
-export const frontendSkills = [
-    "React",
-    "TypeScript",
-    "JavaScript",
-    "TailwindCSS",
-    "React Query",
-    "Redux (RTK)",
-    "SEO",
-    "Next.js",
-    "Framer Motion",
-    "Firebase Auth",
-    "UI/UX Design",
-    "Playwright",
-    "Cypress",
-    "Jest",
-    "Vitest",
-    "HTML",
-    "CSS"
-];
-
-export const backendSkills = [
-    "TypeScript",
-    "Node.js",
-    "Express",
-    "Fastify",
-    "Prisma",
-    "PostgreSQL",
-    "SQLServer",
-    "MySQL",
-    "SQLite",
-    "MongoDB",
-    "REST API",
-    "GraphQL",
-    "JWT",
-    "Firebase Auth",
-    "Hoppscotch",
-    "Postman"
-];
-
-export const devOpsSkills = [
-    "Docker",
-    "Linux",
-    "CI/CD",
-    "GitHub",
-    "GitLab",
-    // "Kubernetes",
-    "NGINX",
-    "Bash",
-    "PowerShell",
-    "Python",
-    "Security",
-    "AWS",
-    "Firebase",
-    "Google Cloud",
-    "Oracle Cloud",
-    "Server management",
-    "Domain & SSL management"
-];
-
-export const otherSkills = [
-    "Agile",
-    "Scrum",
-    "Figma",
-    "Leadership",
-    "Communication",
-    "Building PCs",
-    "C++",
-    "C#",
-    "Rust",
-    "Java",
-    "Flutter",
-    "Dart",
-    "Troubleshooting",
-    "Team Management",
-    "Jetpack Compose",
-    "React Native"
-];
