+++ nextjs/src/services/BlogRequests.ts
@@ -1,5 +1,5 @@
 import { createDirectus, rest, readItem, authentication, readUser, readItems } from "@directus/sdk";
-import { Blog, Project, User } from "@/types";
+import { Blog, User } from "@/types";
 
 // create a Directus client to connect to the Directus API
 const client = createDirectus(process.env.DIRECTUS_URL as string)
@@ -19,6 +19,7 @@ export const fetchBlogs = async () => {
             );
         return sortedBlogs;
     } catch (error) {
+        console.error("Error while fetching blogs", error);
         return [];
     }
 };
@@ -30,6 +31,7 @@ export const fetchBlogById = async (id: string) => {
         if (!res.published) return null;
         return res;
     } catch (error) {
+        console.error("Error while fetching blog", error);
         return null;
     }
 };
@@ -40,6 +42,7 @@ export const getBlogAuthor = async (id?: string) => {
         const res = (await client.request(readUser(id))) as User;
         return res;
     } catch (error) {
+        console.error("Error while fetching blog author", error);
         return null;
     }
 };
@@ -55,6 +58,7 @@ export const getFeaturedBlogs = async () => {
             .slice(0, 3);
         return featuredBlogs;
     } catch (error) {
+        console.error("Error while fetching featured blogs", error);
         return [];
     }
 };
@@ -69,20 +73,7 @@ export const getLatestBlog = async () => {
             .slice(0, 1);
         return latestBlog;
     } catch (error) {
-        return [];
-    }
-};
-
-export const getProjects = async () => {
-    try {
-        const res = (await client.request(
-            readItems("project", {
-                fields: ["*", { image: ["*"], content: ["*"] }]
-            })
-        )) as Project[];
-        const sortedProjects = res.filter((project) => project.status === "published");
-        return sortedProjects;
-    } catch (error) {
+        console.error("Error while fetching latest blog", error);
         return [];
     }
 };
